+++ Classes/Command/AbstractCommandController.php
@@ -3,6 +3,8 @@ namespace T3v\T3vCore\Command;
 
 use \TYPO3\CMS\Extbase\Mvc\Controller\CommandController;
 
+use Colors\Color;
+
 /**
  * Abstract Command Controller Class
  *
@@ -13,11 +15,32 @@ abstract class AbstractCommandController extends CommandController {
    * Helper function to log a message.
    *
    * @param string $message The message
+   * @param string $color The optional color or status, defaults to `white`
    * @return void
    */
-  protected function log($message) {
+  protected function log($message, $color = 'white') {
     if ($message) {
-      echo("{$message}\n");
+      $colorize = new Color();
+
+      switch(true) {
+        case ($color === 'info' || $color === 'blue'):
+          echo $colorize($message)->blue() . PHP_EOL;
+
+          break;
+
+        case ($color === 'warning' || $color === 'yellow'):
+          echo $colorize($message)->yellow() . PHP_EOL;
+
+          break;
+
+        case ($color === 'ok' || $color === 'green'):
+          echo $colorize($message)->green() . PHP_EOL;
+
+          break;
+
+        default:
+          echo $colorize($message)->white() . PHP_EOL;
+      }
     }
   }
 }