+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
+* Colorized logging in `AbstractCommandController`
 * Added `kevinlebrun/colors.php` as Composer dependency
 * Added `AbstractCommandController`